Add `chrono` as an optional crate to replace `time` and `msdos_time`
The `time` crate is no longer actively maintained.  It lists `chrono` as an
alternative, and looking at the 'Recent' statistics on crates.io `chrono`, seems
to be about as popular as `time`.  For all the `chrono` users out there it would
be nice, not to have to convert between `time`'s structs and `chrono`'s structs
whenever working with zip files.

The disadvantage of the PR is that it makes the code base quite a bit more
complex with a lot of `#[cfg]` attributes.  I am especially unhappy with the
added nested code blocks in `read.rs` required to guard multiple code lines with
a single `#[cfg]` attribute, but I could not find a better solution.  The `cfg!`
macro does not help, because both code paths must be valid with any set of
feature flags.  This does not work out, because the one branch needs `chrono`
types (which are only available with the 'chrono' feature flag) and the other
ones requires `time` types (which are only available if the 'chrono' feature
flag has __not__ been set).

Finally I pulled in the `#[cfg]` attribute for `write::FileOptions::default`, so
that we don't need to define the whole function twice.  Technically this does
not directly have anything to do with the `chrono` addition, but made my life a
bit easier since I didn't have to add the guarded definitions of
`last_modified_time` to both versions of the method.