Poison the archive if reading a central file header fails.
The reader can be misaligned if it encounters a corrupt file header before it
find the desired file.

For example, consider an archive that contains files A, B, C and D, and the user
requests file D (say via `by_index(3)`). Let's say the reader successfully
parses the headers of A and B, but fails when parsing C because it's corrupt.
In this case, we want the user to receive an error, but we also want to ensure
that future calls to `by_index` or `by_name` don't parse A and B's headers
again.

One way to do this is to preserve the end position of B's header so that future
calls to `read_files_till` resume from it. However since we've already parsed
B's header and know C's header failed to parse, we can simply consider all
headers that haven't already been parsed to be unreachable.

This commit marks the archive as poisoned if `read_files_till` fails with any
error, so that only headers which have already been successfully parsed will be
considered when getting files from the archive. If a file is requested that
hasn't been parsed, ie its header lies beyond the known corrupted header,
then the function fails immediately.