Only set permissions on files not symlinks
The previous code was setting the permissions on the symlink, which could fail if the file that the symlink points to doesn't exist yet.

Also in my reading on links it seems hat setting the permissions of a link doesn't make a lot of sense anyway - it's the target file permissions that count.

So this change only sets the permissions on the actual files - not on the symlinks.