Remove dependency on thiserror
On my computer, this halves the total build time from 11.7s to 5.4s, as
well as the total size of the artifacts in target.

This derive macro is nice, but it doesn’t justify the increase in
compilation time for dependent crates imo.

This effectively reverts #135, although using the up to date APIs from
std::error::Error.