Stop using byteorder in the "spec" module
Similar to the previous commit: What byteorder provides here can easily
be done with the standard library.

This moves the helper functions that were added in the previous commit
to "spec" (since "read" already imports "spec", but not the other way)
and converts the code to use these helpers instead of byteorder.

Signed-off-by: Uli Schlachter <psychon@znc.in>